# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sydney Project",
        "description": "Sydney Projecta",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/student/submissions": {
            "get": {
                "tags": [
                    "Student Assignments"
                ],
                "summary": "Get assignments for the logged-in student",
                "description": "Retrieves assignments based on the student's enrolled groups and courses.",
                "operationId": "getStudentAssignments",
                "responses": {
                    "200": {
                        "description": "Assignments retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Assignments retrieved successfully."
                                        },
                                        "assignments": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "course_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "lecturer_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "This is title of the assignment"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "This is  description of the assignment"
                                                    },
                                                    "due_date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2025-04-22"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-04-21T18:35:06.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-04-21T18:35:06.000000Z"
                                                    },
                                                    "links": {
                                                        "type": "string",
                                                        "example": "https:www.k.com"
                                                    },
                                                    "course": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Bachelor in Computer Application"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access. Only students can view assignments.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized access. Only students can view assignments."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No groups, courses, or lecturers found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "You are not assigned to any groups yet."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error while retrieving assignments.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred while retrieving assignments."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}